requirements.txt
docker-py==1.10.1
docker-pycreds==0.2.0
requests==2.10.0
six==1.10.0
websocket-client==0.37.0
